Factor out MockChain into mock_chain.go
